data: Use January 2025's police data
data.police.uk data is released monthly on a three-month
delay. This commit updates the link to the data and
where its referenced in the importer.

In an ideal world, this would be automatic, but time
constraints prevent that.

Chosen to limit the time period to the previous six months.
Any time more than that probably one for a specialist database
on account of how large the Spatialite database would be.

Add details on the Crime, Outcome, Stop and Search dataset to
the README.